MB-65976 Handle OR clause when deriving IS NOT NULL predicates
In the problematic query there is an OR clause as a join predicate
and each arm of the OR references a common column that is a
leading index key. The current mechanism for deriving IS NOT NULL
predicate only handles simple join expression in the ON clause.
As a result when join enumeration is in play we did not derive a
IS NOT NULL predicate and thus it did not use the index.

The fix is to add logic to handle OR clause as well when checking
join predicate in the ON clause. As long as a common term exists
in all OR subterms we can use that term to derive a predicate.

Change-Id: I074726cb7dfd784e3c4641584100f38d2907180a
Reviewed-on: https://review.couchbase.org/c/query/+/225485
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Well-Formed: Restriction Checker