MB-65954 Minor adjustment
I miscounted the tokens for the CORRELATED subquery case, the little
block of code in between the CORRELATED keyword and the LPAREN is
counted as well. Thus revert back part of the changes I made to the
error context (the original $3 is correct, I changed to $2 which is
incorrect).

Change-Id: I18b3331c3fb11b0d4fe820958ae21b13d852479a
Reviewed-on: https://review.couchbase.org/c/query/+/225576
Reviewed-by: Vijay Sarathy <vijay.sarathy@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker