MB-65722: Modify timestamp generation in AWR snapshot document key
In the existing implementation the timestamp was the end time of the
previous collection interval that actually inserted workload data.

But this can be misleading in scenarios where there were no
statements recorded by AWR for multiple collection intervals at a
continious stretch. The timestamp would be a timestamp from the last
interval that actually inserted a workload document, which could have
been a long time ago. This can be misleading, especially
during reporting.

With the changes in this patch, the timestamp in the AWR doc key
will be the start time of interval that the AWR document represents.

Change-Id: I97567d56418ffdb3c3f43f750c1ed5c1067cb4ee
Reviewed-on: https://review.couchbase.org/c/query/+/225804
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Vijay Sarathy <vijay.sarathy@couchbase.com>