MB-66076: Track output of expression evaluation correctly.
For correlated subqueries, the output of expression evaluation is tracked as part of (repeated) execution of the subquery.
The current memory quota tracking infrastructure incorrectly tracks the values again after the subquery has been evaluated resulting in memory exceeded quota errors.

We fix this by not tracking these values again for correlated subqueries.

Change-Id: Ie77cbfa205b0881a415fc5b2957245ada1837888
Reviewed-on: https://review.couchbase.org/c/query/+/225842
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>