MB-66147 Avoid formalization of subqueries inside inline UDF
The previous fix for MB-65954 introduced a new formalization step
when an uncorrelated subquery expression is being parsed. The
issue with this bug is that when the subquery is inside inline
UDF, the subquery expression may reference function arguments. In
recent releases (e.g 7.2.7, 7.6.6) if a subquery inside inline
UDF references function arguments we mark the subquery as
correlated, however in earlier releases (e.g. 7.1.x) this is not
done. As a result, if an upgrade is performed from 7.1.x to
7.6.6, since the subquery is not marked as correlated, we go
through the newly added formalization step for uncorrelated
subqueries, and due to references to function arguments the
formalization fails, resulting in error when trying to run the
inline UDF after upgrade.

To fix this, we add a new indicator on the lexer such that we
know if an expression parsing is being done for inline UDF
expressions. We modify the previous fix to skip uncorrelated
subqueries inside inline UDF when performing the new
formalization step. Formalization is done separately for
inline UDF expressions, with the knowlege of function arguments
such that we can properly recognize the references to function
arguments.

Change-Id: I30001f91d31d6231ad94293b833941bac502f66c
Reviewed-on: https://review.couchbase.org/c/query/+/225852
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>