MB-66160 Do not mark invalid index hints as hint error
If an INDEX hint is using an invalid (non-existing) index, then
we immediately mark the hint as ERROR state. However we currently
also set an INDEX_HINT_ERROR flag on the baseKeyspace. In this
bug we have an OR clause, and when the entire OR clause is used
for index selection, we end up with an index hint error; then
when we use individual OR terms to generate a UnionScan, since
the index hint is already marked as error we did not consider
the index hint and thus no hint errors result. This then results
in favoring the UnionScan plan over a single index plan since
one is marked with hint error and the other is not.

The index hint error marking on baseKeyspace is really meant for
situations where a valid index hint is used but cannot be
followed (e.g. lacking appropriate predicate). Thus remove the
marking of index hint error on baseKeyspace if the index is
invalid to start with.

Change-Id: I0ae4353c067019a72d2035249d659d4d7810ddc5
Reviewed-on: https://review.couchbase.org/c/query/+/225892
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>