MB-66176: [BP MB 61782] Update TLS configuration of Query's secure...
ports without closing and re-starting the secure listeners.

A callback is now set in the GetConfigForClient() method of the
listener's TLS configuration.

This callback returns the TLS config to be used for the TLS
handshake when when a ClientHello message is received.

Query will maintain a single TLS configuration for the encrypted port.
This configuration object will be returned by the callback mentioned
above.

Hence, any update to the port's TLS Configuration config returned by
the callback.

Change-Id: Iadc2b505314a6de80e10e6287580ae655110492e
Reviewed-on: https://review.couchbase.org/c/query/+/226058
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>