MB-66219 Proper verification of query plan for queries with a WITH clause
When a prepared statement is executed we first check whether the
query plan is still valid, i.e. has there been any changes in the
underlying keyspaces (e.g. index used in the plan is dropped). If
so we reprepare the plan.

This verification process currently does not work properly if a
query has WITH clauses defined. To fix this we need to traverse
the child operator of the WITH clause when verifying query plan.

Change-Id: Iacafc66e1004701c7a95b9d25f3ca4a570eef256
Reviewed-on: https://review.couchbase.org/c/query/+/226092
Well-Formed: Restriction Checker
Reviewed-by: Vijay Sarathy <vijay.sarathy@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
(cherry picked from commit 716c4425db69d3ada1604a446d7cbaff506e391a)
Reviewed-on: https://review.couchbase.org/c/query/+/226095