MB-60680 Modify index scan cost when LIMIT/OFFSET is available
Currently for an index scan, if LIMIT/OFFSET pushdown can be done,
we first calculate the index scan cost for the entire index scan,
then we take into consideration of LIMIT/OFFSET pushdown in a
separate step. This works ok for composite index scans, but is
more difficult to do the same for Bhive indexes.

We now change the way index scan cost is done when LIMIT/OFFSET
pushdown is available, instead of doing this in two separate
steps, we do all of the costing in one step. For composite
index scan this is fairly straightforward, we just sort of do
whatever was in the second step at the end of the first step.

The change for LIMIT/OFFSET pushdown for Bhive indexes is
still pending.

Change-Id: Ic5bfbf5f1eadbcfff623a3053ebd61f04e1141c9
Reviewed-on: https://review.couchbase.org/c/query/+/226592
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>