MB-66365 Proper handling of histogram keys with escape character during restore
During backup/restore, the content of the system collection is handled
separately. Specifically for CBO stats, we backup/restore each
individual document separately.

The issue with this bug is that if the document is for a histogram, and
the histogram key happens to contain double quotes, then escape
character is added to escape the quotes during backup. When we restore,
we do not properly handle the escaped quotes in the document key
currently since we only strip off the quotes at the beginning and the
ending of the bytes, leaving the escaped character in the string. If
this process is repeated, i.e. repeasted backup/restore, then each time
new escape characters are added to the document key, eventually leading
to the document key exceeding the limit.

To fix this, we now use json.Unmarshal to handle CBO stats document keys
during restore, which handles the escape character properly. In
addition, we also check whether the specific pattern of escaped quotes
exists in the document key (from previous incorrect backup/restore),
and if so, replace the specific pattern with just the regular quotes to
return it to the proper state.

This issue is somewhat specific to CBO stats since we don't expect
escaped characters in either UDF or sequences backups.

Change-Id: I6e85e7bd4d9efca2890c7219f6c5f141d3754cd4
Reviewed-on: https://review.couchbase.org/c/query/+/226860
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>