MB-66154: Fix some admin endpoints not authenticating client certificates
This fixes the following:

* Fixes an issue where some admin endpoints only accepted
Basic Authentication and did not support authentication using client
certificates

* When n2n encryption and client certificate authentication is enabled,
 requests between Query nodes must authenticate using an internal
client certificate.

However, the implementation in MB-52102 mistakenly adds both Basic Auth
header and an internal client certificate in the request.

Modify this to only add the internal client certificate in the request.

Earlier, to prevent flooding the audit logs with requests sent from the
internal Query user, the admin endpoints extracted the username from the
request and checked if the username was "@cbq-engine". And skipped
auditing such requests.

But, the username in the internal client certificate is @internal and
not @cbq-engine. Hence, to identify if a request made is from the Query
service user the following check is made now:

a. User agent is “query”
AND
b. Username starts with “@”.

Change-Id: I60d379e1a346e3e0afa88d203a606477a3a795eb
Reviewed-on: https://review.couchbase.org/c/query/+/226971
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>