MB-66046 Do not perform early order on vector index key ...    ... if rerank is requested
If rerank is requested in the approx_vector_distance function in the
ORDER BY clause, but LIMIT cannot be pushed down, then we should not
be performing early order with projected vector_distance from the
index, since in this case rerank is not performed in the indexer.
Add an extra check during determination of early order to skip the
vector distance function if rerank is being requested.

Change-Id: Iabf0166bd19f67265f2bca285175f294482ba2d6
Reviewed-on: https://review.couchbase.org/c/query/+/226981
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>