MB-70633 Implement conversational support...
... for USING AI statement.

The changes propose to enhance the existing
natural language request processing from
one shot requests, i.e plain USING AI statements
to conversation history aware USING AI statements.

Two new statements are introduced -
1. BEGIN CHAT
   required parameters:
   natural_context

   This statement will create a new chat session
   which can be used only by the user that made
   the request to create it.
   Every chat is uniquely identified with a chatid.
   The response of this request will include the chatid
   for the newly created chat session.

2. END CHAT
   required parameters:
   natural_chatid

   This statement searches for the chat with the chatid
   provided via natural_chatid query parameter and then
   deletes the chat session if the requesting user matches
   the chat user.

USING AI requests when sent with natural_chatid indicate that
the request is part of a chat session. The request's prompt
that is sent to the llm api would include past messages, both
from the user and the assistant. The response from the llm is
saved in the history messages as context for any new request
that may come in.

Monitoring chat sessions:
A new system keyspace, natural_chats stores information
about on going chat sessions. The keyspace supports SELECT
and DELETE. For this to work, admin/natural_chats index handler
and chat handler are introduced. Before processing a chat for retrieval
or deletion a check is made to ensure that the user matches the
user that created the chat session.

The current implementation only guarantees the availabilty of a
chat session in the node where it was created.

Chats are stored in memory using the GenCache struct.
A guard of 2*numcpus limit on the number of ongoing chat sessions
is imposed per node.

For memory reasons, size the prompt for a chat USING AI request is
limited to 1MiB. An error is returned once a chat's prompt out grows
this limit.

Change-Id: Ieeeb5db2148b5a1bb9e90514be67a57014787edc
Reviewed-on: https://review.couchbase.org/c/query/+/227034
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Gaurav J <gaurav.jayaraj@couchbase.com>