MB-66399: Fix ADVISE not providing index recommendations...
for a subset of subqueries.

This issue occrs when the query contains nested subqueries.

The problem is while identifying subqueries to advise, the map that
stores the identified subqueries, is reset everytime a new subquery
is recursively explored.

As a result, some subqueries in the query are not advised.

Change-Id: I754025558fed88ff8c370fc58a55f3424165a296
Reviewed-on: https://review.couchbase.org/c/query/+/227036
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Well-Formed: Restriction Checker