MB-60680 Better handling of rerank in CBO costing
If rerank is requested, but the index does not handle reranking,
query expands the (limit+offset) that is being pushed to the
indexer by a factor, and perform subsequent reranking after
fetchig the documents and calculate vector distance with the
actual vector.

Modify CBO costing to also take into consideration this expanded
(limit+offset) during optimizer costing.

Change-Id: I1003ec7a649e17b51a800ee133edde0dc18b5fa9
Reviewed-on: https://review.couchbase.org/c/query/+/227050
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>