MB-65619 Do not use default selectivity for simple field references ...   ... with CBO if histogram is missing
When customers upgrade to 7.6, since we automatically create optimizer
statistics on a new CREATE INDEX/BUILD INDEX command, it is possible to
now have histograms for a few fields (the new index keys) but missing
histograms for other fields. In such cases currently CBO is used and
for the predicates involving fields without histograms a default
selectivity is used. This default selectivity can be orders of magnitude
different than the actual selectivity, resulting in bad optimizer
choices.

We now detect such situations and instead of using a default selectivity
we just bail out and fall back to RBO in such cases. This should reduce
the chances of plan regression for customers migrating to 7.6.

This new behavior is under the control of a new bit in n1ql_feat_ctrl,
such that for customers already on 7.6 and expecting the behavior with
default selectivity, this new bit can be set. By default the new
behavior, i.e. not using default selectivity is used.

Note that there are also situations where the optimizer uses default
selectivity when an accurate selectivity cannot be properly calculated.
For example, if the predicate is complex, e.g. involving functions,
or for predicates involving unnest aliases, etc. In such situations
we still allow default selectivity to be used, otherwise CBO will not
work for such queries. We only disallow default selectivity for simple
predicate involving simple field references on base tables.

Change-Id: I68100064619fbeec012d6cfb9ab8d77cb8120537
Reviewed-on: https://review.couchbase.org/c/query/+/227129
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>