MB-60680 Fix issue with cardinality estimate for Bhive indexes
CBO has a mechanism to mark filters used to generate index spans on
index keys as "already applied", such that the selectivity of these
filters are not re-applied later on. However, this mechanism does
not currently work for Bhive index if predicates exist for include
columns.

Extend the current mechansim to handle filters on include columns
as well:

 - when marking filters with special flags, add both index spans
   and include spans such that both will be considered the same
   way
 - in addition to keep track of exact filters used for generating
   index spans, also keep track of exact filters used for
   generating include spans, and skip these filters when calculating
   selectivity of post-index filters (we need to generate filters
   on include columns as post-index filters such that we can pass
   them to indexer)

Change-Id: Id4ff48e625a5a3ed8770a963dcf7be4a2428832d
Reviewed-on: https://review.couchbase.org/c/query/+/227244
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>