MB-66573 Do not include bits that's only relevant for higher code lines
When settings change we report the change in query log. For the settings
for n1ql_feat_ctrl, we also report the corresponding feature change
(enabled/disabled). There are however bit definitions in n1ql_feat_ctrl
that is only relevant in higher code lines (e.g introduced for new
capabilities in Morpheus). For such bit changes we should not report
the normal changed feature since the feature does not exist in 7.6.

Modify the logic for n1ql_feat_ctrl to mark bits for higher code lines
differently, and report the change in such bits appropriately when
changed.

Change-Id: I2a072a2f1958665ef01e0a1d870415ecb41b2bce
Reviewed-on: https://review.couchbase.org/c/query/+/227292
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>