MB-66570 Pass prepare context to index advisor
The index advisor is somewhat strange in that it is called from
the planner, but it also uses execution context (and datastore
QueryContext) in most of its processing.

The fix for MB-65619 adds requirement for using planner
PrepareContext as well (for checking bits in n1ql_feat_ctrl),
and this causes panics in the index advisor code since currently
it passes nil as pointer for prepare context (optuil.Context)
when it needs to get selectivity estimates (for cost-based
advise).

To fix this, we now also pass the prepare context from the
planner when performing index advisor tasks.

Change-Id: I8dc4726d3f790397e2ea8addc3fcc50f4b56cf30
Reviewed-on: https://review.couchbase.org/c/query/+/227293
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker