MB-66399: [neo] Fix ADVISE not providing index recommendations...
for a subset of subqueries.

This issue occrs when the query contains nested subqueries.

The problem is while identifying subqueries to advise, the map that
stores the identified subqueries, is reset everytime a new subquery
is recursively explored.

As a result, some subqueries in the query are not advised.

Change-Id: Ic9f898ff4de575d8c3d1f9fc95d032396eb7bf95
Reviewed-on: https://review.couchbase.org/c/query/+/227334
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>