MB-66635 Don't always disregard join filters in cost-based advise
The rule-based and cost-based advise behave differently when a join
query is being advised. The rule-based advise considers join filters
separately than table filters, while the cost-based advise considers
them together in a way, since the cost-based advise uses selectivities
to order index keys. Previously in MB-41263, in order to avoid
generating duplicated index keys with cost-based advise, we always
remove join filters from consideration, with the assumption that
we probably already generated a derived filter with the same
expression as a table filter, and thus considering join filters will
likely generate the same index key twice.

However, we don't always derive IS NOT NULL filter from the join
filter. The derivation is only done if there is an existing index
with that expression as the leading index key. As a result, for this
bug, we did not derive the IS NOT NULL filter, and since we also
did not consider join filters, we failed to generate a proper covering
index (that includes the join column).

To fix this, we modify the previous fix for MB-41263, instead of
not considering join filters, we first attempt to remove all index
keys in join filters that are already referenced else where. This
way we can avoid generating duplicated index keys, but also if there
is no duplication, we still retain the join filter as an index key.

Change-Id: Ib7938e542e08ec44b2f5631b00c6b6521240c10a
Reviewed-on: https://review.couchbase.org/c/query/+/227497
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>