MB-66635 Avoid duplicated keyspace aliases in advise
The output of an Advise statement includes the current indexes
defined, and in some cases we have the keyspace_alias field
containing duplicated entries for "current_indexes".

This is a side-effect of the fix for MB-35117 where we consolidate
identical indexes in "current_indexes" section, however, the alias
is being added regardless of whether it already exists.

Add logic to first check for identical aliases before adding it.

Change-Id: I1126676ec69b8f694cc76b26ad701902d0d87c48
Reviewed-on: https://review.couchbase.org/c/query/+/227562
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>