MB-66614 Delay unmarshal of value for early projection
During Fetch, if early projection is feasible, currently we
search for the top-level field and get the corresponding value
for the field. The issue is if the top-level field is complex,
in case of this bug, it is a nested array field, then the
construction of the value for the projected field can take
time, and this is not desirable in the Fetch operator since
Fetch is already heavily loaded.

Modify the logic such that we don't attempt to unmarshal the
entire value in early projection, we use a parsed value
which makes it much easier to deal with in the Fetch operator.
The actual unmarshal happens later in the plan when the value
is actually used, but hopefully that is in a less-busy
operator.

Change-Id: I8c3fac7bb9b94d02210d564c0dd83ff63a7a8d54
Reviewed-on: https://review.couchbase.org/c/query/+/227577
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker