MB-66657 Avoid delay in migration if there is no buckets
When query service starts up it first checks whether migration is complete,
if not, it attempts to perform migration, which includes a mandatory sleep
of 30 seconds at the beginning. The intent was to avoid hamming KV
immediately after startup. However, for a new installation when there are
no buckets, we still go through this process which includes the 30-second
sleep time.

We now skip the 30-second sleep if there are no buckets to be migrated.

Change-Id: If05d3c9ac1db783ec75c2fbf5e98470597a8536c
Reviewed-on: https://review.couchbase.org/c/query/+/227595
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker