MB-66660 Handle formalization in subquery expression parsing better
The previous fix for MB-65954 added a formalization step when performing
expression parsing on non-correlated subqueries. This addition is
somewhat similar to what we already do for correlated subqueries where
we attempt to find all correlated references. There are several special
handling in case of the correlated subquery since we are only doing
formalization on the subquery itself, we don't have information on
what's above the subquery. As such the formalizer in case of the
correlated subquery is marked with a special flag and there are
several special handling based on this flag.

The newly added formalization step for non-correlated subquery
currently does not have any special flag and as a result it behaves
just like a regular formalization (when the entire query is available).
This could be problematic in certain situations as well, for similar
reason that we are doing formalization on the subquery only and we
don't have information on what's above the subquery. In case of
non-correlated subqueries, what is missing is WITH clause aliases
where a normal formalizer would know about (when the entire query
is being formalized). As such if there is a reference to one of the
WITH aliases, since we cannot properly determine it is a WITH alias,
we incorrectly assign it as a keyspace reference which is incorrect.

To fix this, we also add a special flag to the formalizer in case
of parsing non-correlated subqueries, and add special handling in
a few places (somewhat similar to what we do already for correlated
subqueries). The one specific for this bug is when we formalize
an ExpressionTerm, skip the logic for trying to override previous
decision (in the parser, based on whether the expr has parentheses)
of whether the expression is a keyspace reference. This should
address the issue in this bug while still maintain the proper
behavior for the previous bug (MB-65954).

Change-Id: I3ffbcb690f8d709cf663e4bc420872c92951fb23
Reviewed-on: https://review.couchbase.org/c/query/+/227594
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/227655