MB-66699 Properly mark WITH and binding variables for prepared statments
A related issue was identified during investigation of MB-66660. The
repro for MB-66660 has a reference to a WITH alias in the FROM clause
of a subquery, and the fix for MB-66660 takes care of that. However if
I now moved the reference to the WITH alias outside of the FROM clause
and into the WHERE clause, we now are having similar issues as MB-65954
in that we do not use covering index scan for the subquery if the
prepared statement is executed on a remote node. The reason being the
WITH alias is not properly marked as a WITH alias, and thus during
covering determination we did not properly skip the WITH alias,
resulting in not able to use covering index scan.

The WITH alias is normally marked as such during formalization of the
entire query where we have proper references to the WITH clause. In
this case though when we parse the subquery text at the remote node,
we do a formalization step (as part of fix for MB-65954), however
during this formalization we do not have proper context of the WITH
alias which was defined in a query above the subquery. To properly fix
this, we need a new mechanism to somehow pass the information of the
WITH alias (and also binding variables) from the top down such that we
can properly mark the WITH alias reference. This needs to be done when
we unmarshal the parent query plan (where the WITH alias is defined).

We now adds a new "plan context" to all plan operators. This plan
context is somewhat similar to the formalizer in that it keeps track
of references to various identifiers including WITH alias, binding
variable, and keyspace references (including KeyspaceTerm,
ExpressionTerm, SubqueryTerm and Unnest alias). The mechanism of the
plan context is also very similar to what formalizer does. The plan
context is used when a prepared query plan is shipped to a remote
node and being unmarshalled. We pass along the plan context as we
regenerate the query plan, operator by oeprator, and along the way we:
  - accumulate references to WITH aliases, binding variables, and
     various keyspace references
  - perform expression "transformation" on the relevant expressions
    on each operator. The transformation is just adding appropriate
    identifier flags to identifiers such that we can properly
    recognize each identifier later on (e.g. when planning a subquery),
    the expression itself does not actually change
  - handle binding variables (push/pop) for certain expressions (ARRAY,
    OBJECT, etc), as well as LET clauses

The new plan context is only used when the query plan is unmarshalled
after being shipped from one node to another, it is not used when just
constructing a regular query plan. When it is being used, what it
gives us is the ability to basically try to perform something similar
to the formalization step from the top down. This way, when a WITH
clause is defined, we keep track of the WITH reference, and when we
encounter a subquery expression while walking the operator tree, we
go ahead and properly mark all WITH aliases within the subquery
expression, such that when that subquery gets planned eventually we
would have the proper marking on the identifiers to enable proper
covering determination.

Change-Id: Ibff9c8a7939f0950a02ded794078ebdda5eba943
Reviewed-on: https://review.couchbase.org/c/query/+/227805
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>