MB-66660 Fix issue with intervening correlated subqueries
An additional issue was identified while trying to backport changes
to 7.2.8. In case of intervening correlated subqueries, i.e. two
levels of subqueries with the inner-most subquery being correlated,
we may not be marking the intermediate subquery as correlated
properly when the plan is being decoded in a remote node, resulting
in potentially incorrect result.

When we perform formalization on a subquery expression, we skip
the subquery if the formalization is done for handling subqueries
during unmarshalling of a query plan, since we know the subquery
has been handled separately by the same process when the nested
subquery is encountered by the parser. However, what we failed to
do is in case the nested subquery is correlated, we need to
propagate up the correlation references from the correlated
subquery even when we don't need to traverse the subquery.

Change-Id: I4dd5c82fb31b507271096929b20ce306faea64da
Reviewed-on: https://review.couchbase.org/c/query/+/227738
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>