MB-66660 Handle formalization in subquery expression parsing better
The previous fix for MB-65954 added a formalization step when performing
expression parsing on non-correlated subqueries. This addition is
somewhat similar to what we already do for correlated subqueries where
we attempt to find all correlated references. There are several special
handling in case of the correlated subquery since we are only doing
formalization on the subquery itself, we don't have information on
what's above the subquery. As such the formalizer in case of the
correlated subquery is marked with a special flag and there are
several special handling based on this flag.

The newly added formalization step for non-correlated subquery
currently does not have any special flag and as a result it behaves
just like a regular formalization (when the entire query is available).
This could be problematic in certain situations as well, for similar
reason that we are doing formalization on the subquery only and we
don't have information on what's above the subquery. In case of
non-correlated subqueries, what is missing is WITH clause aliases
where a normal formalizer would know about (when the entire query
is being formalized). As such if there is a reference to one of the
WITH aliases, since we cannot properly determine it is a WITH alias,
we incorrectly assign it as a keyspace reference which is incorrect.

To fix this, we also add a special flag to the formalizer in case
of parsing non-correlated subqueries, and add special handling in
a few places (somewhat similar to what we do already for correlated
subqueries). The one specific for this bug is when we formalize
an ExpressionTerm, skip the logic for trying to override previous
decision (in the parser, based on whether the expr has parentheses)
of whether the expression is a keyspace reference. This should
address the issue in this bug while still maintain the proper
behavior for the previous bug (MB-65954).

MB-66660 Fix issue with intervening correlated subqueries

An additional issue was identified while trying to backport changes
to 7.2.8. In case of intervening correlated subqueries, i.e. two
levels of subqueries with the inner-most subquery being correlated,
we may not be marking the intermediate subquery as correlated
properly when the plan is being decoded in a remote node, resulting
in potentially incorrect result.

When we perform formalization on a subquery expression, we skip
the subquery if the formalization is done for handling subqueries
during unmarshalling of a query plan, since we know the subquery
has been handled separately by the same process when the nested
subquery is encountered by the parser. However, what we failed to
do is in case the nested subquery is correlated, we need to
propagate up the correlation references from the correlated
subquery even when we don't need to traverse the subquery.

Change-Id: I4dd5c82fb31b507271096929b20ce306faea64da
Reviewed-on: https://review.couchbase.org/c/query/+/227738
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/227779