MB-66758: Fix AUS incorrectly reporting error during Query backup...
and restore in mixed mode clusters and Community Edition clusters.

The issue with mixed mode clusters is specific to clusters that have
nodes with 8.0 version and nodes below 7.6.0 version.

This is because, the 8.0 node will attempt to backup/restore the AUS
settings even though AUS is not supported/initialized for the cluster.
And an error is reported.

AUS is only initialized for Enterprise clusters that have completed
migration of CBO statistics.

To fix this issue, skip the backup/restore of AUS settings if AUS is
not supported/initialized.

Additional minor fixes:
* Reduce delay before running cleanup task.
* Tweak coordination doc key prefix

Change-Id: Id14f729c4eec0b72d139797302f2455718a873fa
Reviewed-on: https://review.couchbase.org/c/query/+/227887
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>