MB-66748 Reduce default batch size for UPDATE STATISTICS
When UPDATE STATISTICS is executed with INDEX ALL option, or with
INDEX option and a large number of indexes are specified, we could
have a large number of terms to gather histograms on. In this case
UPDATE STATISTICS gather the histograms in batches, since each
histogram requires a sort. The timeout for UPDATE STATISTICS
applies to each batch, and thus when a timeout error occurs it
implies that there are too much work being done for the batch.

Since UPDATE STATISTICS should be an infrequent command, to
reduce the chance of a timeout error, we reduce the default
batch size. This may make the whole UPDATE STATISTICS command
take longer but should reduce the chance of hitting timeout
errors. The following changes are done:
 - reduce the default batch_size to 20 if spill-to-disk is
   enabled, and 10 if spill-to-disk is disabled
 - if sample_size is specified, also add logic to adjust
   batch_size based on specified sample_size
 - if multiple batches are required, add logic to check the
   status of the operator context before starting each batch,
   such that if the statement is no longer active (e.g. hit
   request timeout) we don't keep executing.
 - add marking for context used internally for Advise and
   AUS such that the above mentioned check does not apply
   (internal UPDATE STATISTICS does not follow the usual
   operator state used in execution package).

Change-Id: I3966ff414c333033c8f57a88937c45237fe02cfc
Reviewed-on: https://review.couchbase.org/c/query/+/227922
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>