MB-66703 Add plan version and mechanism to force plan reprepare
The issue of this bug is that when a query has ORDER BY and
LIMIT/OFFSET, the Order operator can have embedded Limit/Offset
operator. Previous versions only use this information as a guide
for the heap size used, and the actual OFFSET happens later with
a separate Offset operator. Starting with version 7.6.x we now
actually do the OFFSET operation inside the Order itself and
thus we don't need (or generate) a separate Offset operator
after the Order (see MB-33961/MB-61760). Now in a mixed-mode
cluster if the 7.6 node gets a plan from a 7.2 node we could
get into trouble since the plan has a separate Offset operator
and thus we are effectively doing the OFFSET operation twice.

To fix this, we need to be able to detect such situations and
trigger a reprepare of the plan. To make this work:
 - add "plan version" to a prepared statement
   we already have plan version in each prepared statement
   shipped, currently this is a separate number that we keep
   track of, for the purpose of determining whether the plan
   we got comes from a higher version server. We now put the
   plan version as part of the prepared statement itself. This
   is required for the verification step below.
 - add new verification logic for the Order operator (there is
   no previous verification logic for Order since there is no
   "base" object, such as keyspace, index, etc to be verified).
   The new logic checks whether an OFFSET is being done, and
   if so, whether the plan version (in prepared statement) is
   less than the minimum version that has the above-mentioned
   behavioral change for OFFSET handling, if so, "fail" the
   verification, such that a reprepare will be triggered.
 - the new plan version is not included when marshalling the
   prepared statement, since this is more of an internal
   version. It'll be included the same way it currently is,
   i.e., handled separately
 - for system:prepareds, the new plan version is added as
   part of meta data, i.e. can be retrieved as
   'meta().planVersion' if required, this is similar to how
   we currently handle the query plan itself (meta().plan).
 - for subquery plans that uses the prepared mechanism, use
   a dummy plan version (-1) such that we don't fail
   verification for a dummy plan version.

Change-Id: Iee8fb22e30712f0973f566fcf376e961880d8fa6
Reviewed-on: https://review.couchbase.org/c/query/+/228244
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>