MB-66955 Do not transform IS_OBJECT to itself
The planner performs certain transformations from the original query
predicate to facilitate index selection. For Is_Object(input) function,
it currently transform it into (input > EMPTY_OBJECT AND Is_Object(input)),
i.e. it includes the original expression as part of the transformed
AND subterm. This was originally done as part of the fix for MB-19443
in 4.50 time frame. The intention of this transformation is not
completely clear, one can guess that it is done to avoid including
binary documents, however such transformation is kind of dangerous
and may cause infinite recursion if we need to traverse + transform
the predicate, such as when we try to classify expressions.

To fix this, we remove the second part of the transformed AND clause,
instead, just use (input > EMPTY_OBJECT), which can then be used for
index selection purpose. The original predicate will be applied later
in the query plan which then avoids wrong result. This is somewhat
similar to how LIKE predicate is handled in terms of using
transformed predicate for index selection and apply the original
predicate later on.

Change-Id: Ia4b3b3478361c6e611eb5ebe95f59768f9a89dbc
Reviewed-on: https://review.couchbase.org/c/query/+/228356
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>