MB-67035 Unset primary join flag on KeyspaceTerm during join enumeration
When the planner considers join methods for ANSI JOIN, there is a
special case if the join condition is on META().id from one side
of the join. In such cases the KeyspaceTerm is marked with a special
flag (PrimaryJoin) such that if no join method can be found, it'll
fall back to the legacy lookup join as a last resort.

With RBO (or when join enumeration is disabled), we go through each
keyspace once, and set the PrimaryJoin flag when appropriate. However,
with join enumeration, there is a potential issue since we consider
joins with each keyspace multiple times, and currently we do not
unmark the PrimaryJoin flag, which may result in later join
considerations on a KeyspaceTerm incorrectly treating it as a
primary join even if the join condition in the current join being
considered does not include META().id, which results in subsequent
issues since we don't have a valid primary join key.

The fix is to add extra logic in join enumeration, such that after
each nested-loop join consideration, we check and unset PrimaryJoin
flag if found. Similar flag on the filters are also unset.

Change-Id: I5ca4e67dbc5d239ebd00693140a5aadbca34ee08
Reviewed-on: https://review.couchbase.org/c/query/+/228699
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker